% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidean Algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A numerical value}

\item{y}{A numerical value}
}
\value{
The greatest common divisior of two numbers
}
\description{
The Euclidean Algorithm
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
