% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph containing the edges and the weight of the edges.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{edges}
  \item{v2}{edges}
  \item{w}{weight of the edges}
}}
\usage{
wiki_graph
}
\description{
A data frame containing three variables; the edges of a graph and 
the weight of the edges.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstras_algorithm}
}
\keyword{datasets}
